huge refactor , refactored the logic into small pure functions
