--- src/store/models.js
@@ -1,4 +1,0 @@
-import timeLine from './timeLine'
-export default {
-    timeLine
-}
