+++ src/store/models/models.index.js
@@ -1,0 +1,4 @@
+import timeline from './timeline/timeline.model'
+export const models =  {
+    timeline
+}
