+++ src/store/store.index.js
@@ -1,5 +1,5 @@
 import { init } from "@rematch/core";
-import * as models from "./models";
+import {models} from "./models/models.index";
 const store = init({ models });
 export default store;
