+++ src/utils/canvasUtils.js
